remove splitEntityId duplicate
