+++ apps/chat/src/components/Chat/Publish/PublicationItemsList.tsx
@@ -6,7 +6,7 @@ import classNames from 'classnames';
 import { useTranslation } from '@/src/hooks/useTranslation';
 
 import { constructPath } from '@/src/utils/app/file';
-import { splitEntityId } from '@/src/utils/app/folders';
+import { splitEntityId } from '@/src/utils/app/shared-utils';
 
 import { Conversation } from '@/src/types/chat';
 import { FeatureType } from '@/src/types/common';
