+++ apps/chat/src/store/share/share.epics.ts
@@ -27,7 +27,6 @@ import {
   isAttachmentLink,
   isConversationHasExternalAttachments,
 } from '@/src/utils/app/file';
-import { splitEntityId } from '@/src/utils/app/folders';
 import {
   isApplicationId,
   isConversationId,
@@ -38,6 +37,7 @@ import {
 import { EnumMapper } from '@/src/utils/app/mappers';
 import { isEntityIdPublic } from '@/src/utils/app/publications';
 import { hasWritePermission } from '@/src/utils/app/share';
+import { splitEntityId } from '@/src/utils/app/shared-utils';
 import { translate } from '@/src/utils/app/translation';
 import { ApiUtils, parseConversationApiKey } from '@/src/utils/server/api';
