+++ apps/chat/src/utils/app/folders.ts
@@ -14,11 +14,10 @@ import { Prompt } from '@/src/types/prompt';
 import { EntityFilters } from '@/src/types/search';
 
 import { DEFAULT_FOLDER_NAME } from '@/src/constants/default-ui-settings';
-import { ROOT_SECTION_NAME } from '@/src/constants/sections';
 
 import { doesHaveDotsInTheEnd, prepareEntityName } from './common';
 import { hasWritePermission } from './share';
-import { isReplayConversation } from './shared-utils';
+import { isReplayConversation, splitEntityId } from './shared-utils';
 
 import {
   Attachment,
@@ -368,35 +367,6 @@ export const addGeneratedFolderId = (
     };
   }
   return folder as FolderInterface;
-};
-
-// {apikey}/{bucket}/path.../name
-export const splitEntityId = (
-  id: string,
-): {
-  bucket: string;
-  name: string;
-  parentPath: string | undefined;
-  apiKey: string;
-  isRoot: boolean;
-} => {
-  const parts = id.split('/');
-  const parentPath =
-    parts.length > 3
-      ? constructPath(...parts.slice(2, parts.length - 1))
-      : undefined;
-
-  const isRoot = parts.length < 3;
-
-  const name = isRoot ? ROOT_SECTION_NAME : parts[parts.length - 1];
-
-  return {
-    apiKey: parts[0],
-    bucket: parts[1],
-    parentPath,
-    name,
-    isRoot,
-  };
 };
 
 export const getParentFolderIdsFromFolderId = (path?: string): string[] => {
